Move i2c to its own folder
